requirements.txt
geonamescache
pandas
numpy
pytesseract
imutils
matplotlib
Pillow
requests
playwright==1.11
opencv-python==4.2.0.34
TikTokAPI
scipy